fastapi/requirements.txt
fastapi==0.89.1
psycopg2-binary~=2.9.5
uvicorn==0.20.0